Crate rustpython_vm

This crate contains most python logic.

Compilation

Bytecode

Import mechanics

Base objects

Modules

bytecode
Implement python as a virtual machine with bytecodes. This module implements bytecode structure.

cformat

eval

exceptions

format

function

import

macros

obj
This package contains the python basic/builtin types

py_serde

pyobject

scope

slots

stdlib

types

util

Macros

arg_check

class_or_notimplemented

count_tts

extend_class

extend_module

match_class
Macro to match on the built-in class of a Python object.

no_kwargs

py_class

py_compile_bytecode

py_compile_bytecode

py_module

py_namespace

replace_expr

type_check

Structs

PySettings
Struct containing all kind of settings for the python vm.

VirtualMachine
Top level container of a python virtual machine. In theory you could create more instances of this struct and have them operate fully isolated.

Enums

InitParameter

Attribute Macros

pyclass

pyimpl

pystruct_sequence

Derive Macros

FromArgs