Trait rustpython_vm::function::IntoPyNativeFunc

pub trait IntoPyNativeFunc<T, R> {
    fn into_func(self) -> PyNativeFunc;
}

Implemented by types that are or can generate built-in functions.

For example, any function that:

Accepts a sequence of types that implement FromArgs, followed by a &VirtualMachine

Returns some type that implements IntoPyObject

will generate a PyNativeFunc that performs the appropriate type and arity checking, any requested conversions, and then if successful call the function with the bound values.

A bare PyNativeFunc also implements this trait, allowing the above to be done manually, for rare situations that don't fit into this model.

Required methods

fn into_func(self) -> PyNativeFunc

Loading content...