Module rustpython_vm::obj::objtype

Structs

PyClass
type(object_or_name, bases, dict) type(object) -> the object's type type(name, bases, dict) -> a new type

Functions

isinstance
Determines if obj actually an instance of cls, this doesn't call instancecheck, so only use this if cls is known to have not overridden the base instancecheck magic method.

issubclass
Determines if subclass is actually a subclass of cls, this doesn't call subclasscheck, so only use this if cls is known to have not overridden the base subclasscheck magic method.

new

type_new

Type Definitions

PyClassRef