Function rustpython_vm::obj::objtype::issubclass

pub fn issubclass(subclass: &PyClassRef, cls: &PyClassRef) -> bool

Determines if subclass is actually a subclass of cls, this doesn't call subclasscheck, so only use this if cls is known to have not overridden the base subclasscheck magic method.