Enum rustpython_vm::pyobject::PyObjectKind

pub enum PyObjectKind {
    String {
        value: String,
    },
    Integer {
        value: BigInt,
    },
    Float {
        value: f64,
    },
    Complex {
        value: Complex64,
    },
    Bytes {
        value: Vec<u8>,
    },
    Sequence {
        elements: Vec<PyObjectRef>,
    },
    Dict {
        elements: DictContentType,
    },
    Set {
        elements: HashMap<usize, PyObjectRef>,
    },
    Iterator {
        position: usize,
        iterated_obj: PyObjectRef,
    },
    Slice {
        start: Option<i32>,
        stop: Option<i32>,
        step: Option<i32>,
    },
    Code {
        code: CodeObject,
    },
    Frame {
        frame: Frame,
    },
    Function {
        code: PyObjectRef,
        scope: PyObjectRef,
        defaults: PyObjectRef,
    },
    Generator {
        frame: Frame,
    },
    BoundMethod {
        function: PyObjectRef,
        object: PyObjectRef,
    },
    Scope {
        scope: Scope,
    },
    Module {
        name: String,
        dict: PyObjectRef,
    },
    None,
    Class {
        name: String,
        dict: PyObjectRef,
        mro: Vec<PyObjectRef>,
    },
    WeakRef {
        referent: PyObjectWeakRef,
    },
    Instance {
        dict: PyObjectRef,
    },
    RustFunction {
        function: Box<dyn Fn(&mut VirtualMachine, PyFuncArgs) -> PyResult>,
    },
}

Rather than determining the type of a python object, this enum is more a holder for the rust payload of a python object. It is more a carrier of rust data for a particular python object. Determine the python type by using for example the .typ() method on a python object.

Variants
String
Fields of String

value: String

Integer
Fields of Integer

value: BigInt

Float
Fields of Float

value: f64

Complex
Fields of Complex

value: Complex64

Bytes
Fields of Bytes

value: Vec<u8>

Sequence
Fields of Sequence

elements: Vec<PyObjectRef>

Dict
Fields of Dict

elements: DictContentType

Set
Fields of Set

elements: HashMap<usize, PyObjectRef>

Iterator
Fields of Iterator

position: usize

iterated_obj: PyObjectRef

Slice
Fields of Slice

start: Option<i32>

stop: Option<i32>

step: Option<i32>

Code
Fields of Code

code: CodeObject

Frame
Fields of Frame

frame: Frame

Function
Fields of Function

code: PyObjectRef

scope: PyObjectRef

defaults: PyObjectRef

Generator
Fields of Generator

frame: Frame

BoundMethod
Fields of BoundMethod

function: PyObjectRef

object: PyObjectRef

Scope
Fields of Scope

scope: Scope

Module
Fields of Module

name: String

dict: PyObjectRef

NoneClass
Fields of Class

name: String

dict: PyObjectRef

mro: Vec<PyObjectRef>

WeakRef
Fields of WeakRef

referent: PyObjectWeakRef

Instance
Fields of Instance

dict: PyObjectRef

RustFunction
Fields of RustFunction

function: Box<dyn Fn(&mut VirtualMachine, PyFuncArgs) -> PyResult>