Enum rustpython_vm::pyobject::PyObjectPayload

pub enum PyObjectPayload {
    String {
        value: String,
    },
    Integer {
        value: BigInt,
    },
    Float {
        value: f64,
    },
    Complex {
        value: Complex64,
    },
    Bytes {
        value: Vec<u8>,
    },
    Sequence {
        elements: Vec<PyObjectRef>,
    },
    Dict {
        elements: DictContentType,
    },
    Set {
        elements: HashMap<u64, PyObjectRef>,
    },
    Iterator {
        position: usize,
        iterated_obj: PyObjectRef,
    },
    EnumerateIterator {
        counter: BigInt,
        iterator: PyObjectRef,
    },
    FilterIterator {
        predicate: PyObjectRef,
        iterator: PyObjectRef,
    },
    MapIterator {
        mapper: PyObjectRef,
        iterators: Vec<PyObjectRef>,
    },
    ZipIterator {
        iterators: Vec<PyObjectRef>,
    },
    Slice {
        start: Option<BigInt>,
        stop: Option<BigInt>,
        step: Option<BigInt>,
    },
    Range {
        range: RangeType,
    },
    MemoryView {
        obj: PyObjectRef,
    },
    Code {
        code: CodeObject,
    },
    Frame {
        frame: Frame,
    },
    Function {
        code: PyObjectRef,
        scope: PyObjectRef,
        defaults: PyObjectRef,
    },
    Generator {
        frame: Frame,
    },
    BoundMethod {
        function: PyObjectRef,
        object: PyObjectRef,
    },
    Scope {
        scope: Scope,
    },
    Module {
        name: String,
        dict: PyObjectRef,
    },
    None,
    NotImplemented,
    Class {
        name: String,
        dict: RefCell<PyAttributes>,
        mro: Vec<PyObjectRef>,
    },
    WeakRef {
        referent: PyObjectWeakRef,
    },
    Instance {
        dict: RefCell<PyAttributes>,
    },
    RustFunction {
        function: Box<dyn Fn(&mut VirtualMachine, PyFuncArgs) -> PyResult>,
    },
}

Rather than determining the type of a python object, this enum is more a holder for the rust payload of a python object. It is more a carrier of rust data for a particular python object. Determine the python type by using for example the .typ() method on a python object.

Variants
String
Fields of String

value: String
Integer
Fields of Integer

value: BigInt
Float
Fields of Float

value: f64
Complex
Fields of Complex

value: Complex64
Bytes
Fields of Bytes

value: Vec<u8>
Sequence
Fields of Sequence

elements: Vec<PyObjectRef>
Dict
Fields of Dict

elements: DictContentType
Set
Fields of Set

elements: HashMap<u64, PyObjectRef>
Iterator
Fields of Iterator

position: usizeiterated_obj: PyObjectRef
EnumerateIterator
Fields of EnumerateIterator

counter: BigIntiterator: PyObjectRef
FilterIterator
Fields of FilterIterator

predicate: PyObjectRefiterator: PyObjectRef
MapIterator
Fields of MapIterator

mapper: PyObjectRefiterators: Vec<PyObjectRef>
ZipIterator
Fields of ZipIterator

iterators: Vec<PyObjectRef>
Slice
Fields of Slice

start: Option<BigInt>stop: Option<BigInt>step: Option<BigInt>
Range
Fields of Range

range: RangeType
MemoryView
Fields of MemoryView

obj: PyObjectRef
Code
Fields of Code

code: CodeObject
Frame
Fields of Frame

frame: Frame
Function
Fields of Function

code: PyObjectRefscope: PyObjectRefdefaults: PyObjectRef
Generator
Fields of Generator

frame: Frame
BoundMethod
Fields of BoundMethod

function: PyObjectRefobject: PyObjectRef
Scope
Fields of Scope

scope: Scope
Module
Fields of Module

name: Stringdict: PyObjectRef
NoneNotImplementedClass
Fields of Class

name: Stringdict: RefCell<PyAttributes>mro: Vec<PyObjectRef>
WeakRef
Fields of WeakRef

referent: PyObjectWeakRef
Instance
Fields of Instance

dict: RefCell<PyAttributes>
RustFunction
Fields of RustFunction

function: Box<dyn Fn(&mut VirtualMachine, PyFuncArgs) -> PyResult>