Struct rustpython_vm::pyobject::PyRef

pub struct PyRef<T> { /* fields omitted */ }

A reference to a Python object.

Note that a PyRef<T> can only deref to a shared / immutable reference. It is the payload type's responsibility to handle (possibly concurrent) mutability with locks or concurrent data structures if required.

A PyRef<T> can be directly returned from a built-in function to handle situations (such as when implementing in-place methods such as __iadd__) where a reference to the same object must be returned.

Methods

impl<T: PyValue> PyRef<T>

pub fn as_object(&self) -> &PyObjectRef

pub fn into_object(self) -> PyObjectRef

pub fn typ(&self) -> PyClassRef