Trait rustpython_vm::pyobject::IntoPyObject

pub trait IntoPyObject {
    fn into_pyobject(self, vm: &VirtualMachine) -> PyResult;
}

Implemented by any type that can be returned from a built-in Python function.

IntoPyObject has a blanket implementation for any built-in object payload, and should be implemented by many primitive Rust types, allowing a built-in function to simply return a bool or a usize for example.

Required methods

fn into_pyobject(self, vm: &VirtualMachine) -> PyResult

Loading content...