Struct rustpython_vm::VirtualMachine

pub struct VirtualMachine {
    pub builtins: PyObjectRef,
    pub sys_module: PyObjectRef,
    pub stdlib_inits: RefCell<HashMap<String, StdlibInitFunc>>,
    pub ctx: PyContext,
    pub frames: RefCell<Vec<PyObjectRef>>,
    pub wasm_id: Option<String>,
}

Top level container of a python virtual machine. In theory you could create more instances of this struct and have them operate fully isolated.

Fields
builtins: PyObjectRefsys_module: PyObjectRefstdlib_inits: RefCell<HashMap<String, StdlibInitFunc>>ctx: PyContextframes: RefCell<Vec<PyObjectRef>>wasm_id: Option<String>
Methods

impl VirtualMachine

pub fn new() -> VirtualMachine

Create a new VirtualMachine structure.

pub fn run_code_obj(&self, code: PyObjectRef, scope: Scope) -> PyResult

pub fn run_frame_full(&self, frame: PyObjectRef) -> PyResult

pub fn run_frame(&self, frame: PyObjectRef) -> PyResult<ExecutionResult>

pub fn current_frame(&self) -> Ref<Frame>

pub fn current_scope(&self) -> Ref<Scope>

pub fn class(&self, module: &str, class: &str) -> PyObjectRef

pub fn new_str(&self, s: String) -> PyObjectRef

Create a new python string object.

pub fn new_int<T: Into<BigInt>>(&self, i: T) -> PyObjectRef

Create a new python int object.

pub fn new_bool(&self, b: bool) -> PyObjectRef

Create a new python bool object.

pub fn new_dict(&self) -> PyObjectRef

pub fn new_empty_exception(&self, exc_type: PyObjectRef) -> PyResult

pub fn new_exception(&self, exc_type: PyObjectRef, msg: String) -> PyObjectRef

pub fn new_attribute_error(&self, msg: String) -> PyObjectRef

pub fn new_type_error(&self, msg: String) -> PyObjectRef

pub fn new_unsupported_operand_error(
&self,
a: PyObjectRef,
b: PyObjectRef,
op: &str
) -> PyObjectRef

pub fn new_os_error(&self, msg: String) -> PyObjectRef

pub fn new_value_error(&self, msg: String) -> PyObjectRef

Create a new python ValueError object. Useful for raising errors from python functions implemented in rust.

pub fn new_key_error(&self, msg: String) -> PyObjectRef

pub fn new_index_error(&self, msg: String) -> PyObjectRef

pub fn new_not_implemented_error(&self, msg: String) -> PyObjectRef

pub fn new_zero_division_error(&self, msg: String) -> PyObjectRef

pub fn new_overflow_error(&self, msg: String) -> PyObjectRef

pub fn get_none(&self) -> PyObjectRef

pub fn get_type(&self) -> PyObjectRef

pub fn get_object(&self) -> PyObjectRef

pub fn get_locals(&self) -> PyObjectRef

pub fn context(&self) -> &PyContext

pub fn to_str(&self, obj: &PyObjectRef) -> PyResult<PyStringRef>

pub fn to_pystr(&self, obj: &PyObjectRef) -> Result<String, PyObjectRef>

pub fn to_repr(&self, obj: &PyObjectRef) -> PyResult<PyStringRef>

pub fn import(&self, module: &str) -> PyResult

pub fn isinstance(&self, obj: &PyObjectRef, cls: &PyObjectRef) -> PyResult<bool>

Determines if obj is an instance of cls, either directly, indirectly or virtually via the instancecheck magic method.

pub fn issubclass(
&self,
subclass: &PyObjectRef,
cls: &PyObjectRef
) -> PyResult<bool>

Determines if subclass is a subclass of cls, either directly, indirectly or virtually via the subclasscheck magic method.

pub fn call_get_descriptor(
&self,
attr: PyObjectRef,
obj: PyObjectRef
) -> PyResult

pub fn call_method<T>(
&self,
obj: &PyObjectRef,
method_name: &str,
args: T
) -> PyResult where
T: Into<PyFuncArgs>,

pub fn invoke<T>(&self, func_ref: PyObjectRef, args: T) -> PyResult where
T: Into<PyFuncArgs>,

pub fn invoke_with_locals(
&self,
function: PyObjectRef,
cells: PyObjectRef,
locals: PyObjectRef
) -> PyResult

pub fn extract_elements(
&self,
value: &PyObjectRef
) -> PyResult<Vec<PyObjectRef>>

pub fn get_attribute<T>(&self, obj: PyObjectRef, attr_name: T) -> PyResult where
T: TryIntoRef<PyString>,

pub fn set_attr(
&self,
obj: &PyObjectRef,
attr_name: PyObjectRef,
attr_value: PyObjectRef
) -> PyResult

pub fn del_attr(&self, obj: &PyObjectRef, attr_name: PyObjectRef) -> PyResult

pub fn get_method(&self, obj: PyObjectRef, method_name: &str) -> PyResult

pub fn call_or_unsupported<F>(
&self,
obj: PyObjectRef,
arg: PyObjectRef,
method: &str,
unsupported: F
) -> PyResult where
F: Fn(&VirtualMachine, PyObjectRef, PyObjectRef) -> PyResult,

Calls a method on obj passing arg, if the method exists.

Otherwise, or if the result is the special NotImplemented built-in constant, calls unsupported to determine fallback value.

pub fn call_or_reflection(
&self,
lhs: PyObjectRef,
rhs: PyObjectRef,
default: &str,
reflection: &str,
unsupported: fn(_: &VirtualMachine, _: PyObjectRef, _: PyObjectRef) -> PyResult
) -> PyResult

Calls a method, falling back to its reflection with the operands reversed, and then to the value provided by unsupported.

For example: the following:

call_or_reflection(lhs, rhs, "__and__", "__rand__", unsupported)

Calls __and__ with lhs and rhs.

If above is not implemented, calls __rand__ with rhs and lhs.

If above is not implemented, invokes unsupported for the result.

pub fn serialize(&self, obj: &PyObjectRef) -> PyResult<String>

pub fn deserialize(&self, s: &str) -> PyResult

pub fn _sub(&self, a: PyObjectRef, b: PyObjectRef) -> PyResult

pub fn _isub(&self, a: PyObjectRef, b: PyObjectRef) -> PyResult

pub fn _add(&self, a: PyObjectRef, b: PyObjectRef) -> PyResult

pub fn _iadd(&self, a: PyObjectRef, b: PyObjectRef) -> PyResult

pub fn _mul(&self, a: PyObjectRef, b: PyObjectRef) -> PyResult

pub fn _imul(&self, a: PyObjectRef, b: PyObjectRef) -> PyResult

pub fn _matmul(&self, a: PyObjectRef, b: PyObjectRef) -> PyResult

pub fn _imatmul(&self, a: PyObjectRef, b: PyObjectRef) -> PyResult

pub fn _truediv(&self, a: PyObjectRef, b: PyObjectRef) -> PyResult

pub fn _itruediv(&self, a: PyObjectRef, b: PyObjectRef) -> PyResult

pub fn _floordiv(&self, a: PyObjectRef, b: PyObjectRef) -> PyResult

pub fn _ifloordiv(&self, a: PyObjectRef, b: PyObjectRef) -> PyResult

pub fn _pow(&self, a: PyObjectRef, b: PyObjectRef) -> PyResult

pub fn _ipow(&self, a: PyObjectRef, b: PyObjectRef) -> PyResult

pub fn _mod(&self, a: PyObjectRef, b: PyObjectRef) -> PyResult

pub fn _imod(&self, a: PyObjectRef, b: PyObjectRef) -> PyResult

pub fn _lshift(&self, a: PyObjectRef, b: PyObjectRef) -> PyResult

pub fn _ilshift(&self, a: PyObjectRef, b: PyObjectRef) -> PyResult

pub fn _rshift(&self, a: PyObjectRef, b: PyObjectRef) -> PyResult

pub fn _irshift(&self, a: PyObjectRef, b: PyObjectRef) -> PyResult

pub fn _xor(&self, a: PyObjectRef, b: PyObjectRef) -> PyResult

pub fn _ixor(&self, a: PyObjectRef, b: PyObjectRef) -> PyResult

pub fn _or(&self, a: PyObjectRef, b: PyObjectRef) -> PyResult

pub fn _ior(&self, a: PyObjectRef, b: PyObjectRef) -> PyResult

pub fn _and(&self, a: PyObjectRef, b: PyObjectRef) -> PyResult

pub fn _iand(&self, a: PyObjectRef, b: PyObjectRef) -> PyResult

pub fn _eq(&self, a: PyObjectRef, b: PyObjectRef) -> PyResult

pub fn _ne(&self, a: PyObjectRef, b: PyObjectRef) -> PyResult

pub fn _lt(&self, a: PyObjectRef, b: PyObjectRef) -> PyResult

pub fn _le(&self, a: PyObjectRef, b: PyObjectRef) -> PyResult

pub fn _gt(&self, a: PyObjectRef, b: PyObjectRef) -> PyResult

pub fn _ge(&self, a: PyObjectRef, b: PyObjectRef) -> PyResult