Trait rustpython_vm::function::FromArgs

pub trait FromArgs: Sized {
    fn from_args(
        vm: &VirtualMachine, 
        args: &mut PyFuncArgs
    ) -> Result<Self, ArgumentError>;

    fn arity() -> RangeInclusive<usize> { ... }
}

Implemented by any type that can be accepted as a parameter to a built-in function.

Required methods

fn from_args(
vm: &VirtualMachine,
args: &mut PyFuncArgs
) -> Result<Self, ArgumentError>

Extracts this item from the next argument(s).

Loading content...
Provided methods

fn arity() -> RangeInclusive<usize>

The range of positional arguments permitted by the function signature.

Returns an empty range if not applicable.

Loading content...