Trait rustpython_vm::pyobject::TryFromObject

pub trait TryFromObject: Sized {
    fn try_from_object(vm: &VirtualMachine, obj: PyObjectRef) -> PyResult<Self>;
}

Implemented by any type that can be created from a Python object.

Any type that implements TryFromObject is automatically FromArgs, and so can be accepted as a argument to a built-in function.

Required methods

fn try_from_object(vm: &VirtualMachine, obj: PyObjectRef) -> PyResult<Self>

Attempt to convert a Python object to a value of this type.

Loading content...