Type Definition rustpython_vm::pyobject::PyObjectRef

type PyObjectRef = Rc<PyObject<dyn PyObjectPayload>>;

The PyObjectRef is one of the most used types. It is a reference to a python object. A single python object can have multiple references, and this reference counting is accounted for by this type. Use the .clone() method to create a new reference and increment the amount of references to the python object by 1.