Struct rustpython_vm::VirtualMachine

pub struct VirtualMachine {
    pub sys_module: PyObjectRef,
    pub stdlib_inits: HashMap<String, StdlibInitFunc>,
    pub ctx: PyContext,
    pub current_frame: Option<PyObjectRef>,
    // some fields omitted
}

Top level container of a python virtual machine. In theory you could create more instances of this struct and have them operate fully isolated.

Fields
sys_module: PyObjectRef stdlib_inits: HashMap<String, StdlibInitFunc> ctx: PyContext current_frame: Option<PyObjectRef>
Methods

impl VirtualMachine

pub fn new() -> VirtualMachine

Create a new VirtualMachine structure.

pub fn run_code_obj(
&mut self,
code: PyObjectRef,
scope: PyObjectRef
) -> PyResult

pub fn new_str(&self, s: String) -> PyObjectRef

Create a new python string object.

pub fn new_bool(&self, b: bool) -> PyObjectRef

Create a new python bool object.

pub fn new_dict(&self) -> PyObjectRef

pub fn new_exception(
&mut self,
exc_type: PyObjectRef,
msg: String
) -> PyObjectRef

pub fn new_type_error(&mut self, msg: String) -> PyObjectRef

pub fn new_value_error(&mut self, msg: String) -> PyObjectRef

Create a new python ValueError object. Useful for raising errors from python functions implemented in rust.

pub fn new_not_implemented_error(&mut self, msg: String) -> PyObjectRef

pub fn new_scope(&mut self, parent_scope: Option<PyObjectRef>) -> PyObjectRef

pub fn get_none(&self) -> PyObjectRef

pub fn get_type(&self) -> PyObjectRef

pub fn get_object(&self) -> PyObjectRef

pub fn get_locals(&self) -> PyObjectRef

pub fn context(&self) -> &PyContext

pub fn get_builtin_scope(&mut self) -> PyObjectRef

pub fn to_str(&mut self, obj: &PyObjectRef) -> PyResult

pub fn to_pystr(&mut self, obj: &PyObjectRef) -> Result<String, PyObjectRef>

pub fn to_repr(&mut self, obj: &PyObjectRef) -> PyResult

pub fn call_get_descriptor(
&mut self,
attr: PyObjectRef,
obj: PyObjectRef
) -> PyResult

pub fn call_method(
&mut self,
obj: &PyObjectRef,
method_name: &str,
args: Vec<PyObjectRef>
) -> PyResult

pub fn call_method_pyargs(
&mut self,
obj: &PyObjectRef,
method_name: &str,
args: PyFuncArgs
) -> PyResult

pub fn invoke(&mut self, func_ref: PyObjectRef, args: PyFuncArgs) -> PyResult

pub fn extract_elements(
&mut self,
value: &PyObjectRef
) -> Result<Vec<PyObjectRef>, PyObjectRef>

pub fn get_attribute(
&mut self,
obj: PyObjectRef,
attr_name: PyObjectRef
) -> PyResult

pub fn del_attr(
&mut self,
obj: &PyObjectRef,
attr_name: PyObjectRef
) -> PyResult

pub fn get_method(&mut self, obj: PyObjectRef, method_name: &str) -> PyResult

pub fn call_or_unsupported(
&mut self,
a: PyObjectRef,
b: PyObjectRef,
d: &str,
r: &str,
op: &str
) -> PyResult

Calls default method, reverse method or exception

a - First argument.

b - Second argument.

d - Default method to try and call (such as __and__).

r - Reverse method to try and call (such as __rand__), in case first one fails.

op - Operator for the exception text, for example &.

Given the above example, it will

Try to call __and__ with a and b

If above fails try to call __rand__ with a and b

If above fails throw an exception: TypeError: Unsupported operand types for '&': 'float' and 'int' if a is of type float and b of type int

pub fn _sub(&mut self, a: PyObjectRef, b: PyObjectRef) -> PyResult

pub fn _add(&mut self, a: PyObjectRef, b: PyObjectRef) -> PyResult

pub fn _mul(&mut self, a: PyObjectRef, b: PyObjectRef) -> PyResult

pub fn _div(&mut self, a: PyObjectRef, b: PyObjectRef) -> PyResult

pub fn _pow(&mut self, a: PyObjectRef, b: PyObjectRef) -> PyResult

pub fn _modulo(&mut self, a: PyObjectRef, b: PyObjectRef) -> PyResult

pub fn _xor(&mut self, a: PyObjectRef, b: PyObjectRef) -> PyResult

pub fn _or(&mut self, a: PyObjectRef, b: PyObjectRef) -> PyResult

pub fn _and(&mut self, a: PyObjectRef, b: PyObjectRef) -> PyResult

pub fn _eq(&mut self, a: &PyObjectRef, b: PyObjectRef) -> PyResult

pub fn _ne(&mut self, a: &PyObjectRef, b: PyObjectRef) -> PyResult

pub fn _lt(&mut self, a: &PyObjectRef, b: PyObjectRef) -> PyResult

pub fn _le(&mut self, a: &PyObjectRef, b: PyObjectRef) -> PyResult

pub fn _gt(&mut self, a: &PyObjectRef, b: PyObjectRef) -> PyResult

pub fn _ge(&mut self, a: &PyObjectRef, b: PyObjectRef) -> PyResult